requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
qrcode[pil]==7.4.2
requests==2.31.0
Pillow==10.1.0
aiohttp==3.9.1
Flask==3.0.0
Flask-CORS==4.0.0
